Moved Location of dbConfig to enable Test Cases. Updated Entity and repopulateDatabase.py with the corresponding import change.
