+++ app/entity/HealthStaffUser.py
@@ -1,4 +1,4 @@
-from ...dbConfig import dbConnect, dbDisconnect
+from ..dbConfig import dbConnect, dbDisconnect
 from .User import User
 
 class HealthStaffUser(User):
